+++ js/CU/Agendar.js
@@ -1,28 +1,14 @@
     import {agendas} from "../data.js"
     import {Agenda} from "../clases.js"
+  
 
-    const btnFormulario = document.getElementById("btnFormulario");
-    
-
-
-    btnFormulario.addEventListener("click", () => {
-      const nombre = document.getElementById("name").value;
-      const email = document.getElementById("emails").value;
-      const servicio = document.getElementById("servicio-select").value;
-      const fecha = document.getElementById("date").value;
-      const hora = document.getElementById("inputHora").value;
-      const slcBarberosAgenda = document.querySelector(".barberSelect").value;
-
-      agendarCita(nombre, email, servicio, slcBarberosAgenda, fecha, hora);
-    })
-
-    export default function agendarCita(nombre, email, servicio, slcBarberosAgenda, fecha, hora) {
+    function agendarCita(nombre, email, servicio, slcBarberosAgenda, fecha, hora) {
       
       console.log(nombre, email, servicio, slcBarberosAgenda, fecha, hora)  
 
       if (nombre == "" || email == "" || servicio  == "" || fecha == "" || hora == "") {
-        alert('Hay campos obligatorios');
-        return;
+
+        return 'Campos invalidos';
       }
 
       const nuevaFecha = new Date(fecha)
@@ -30,21 +16,17 @@
       
       const horaTotal = hora.split(":")
       
-      console.log(agendas)
-      
       nuevaFecha.setHours(horaTotal[0])
       nuevaFecha.setMinutes(horaTotal[1])
       
     
     const existe = agendas.some(cita => {
 
-      console.log(cita.barbero == slcBarberosAgenda)  
       return ((cita.barbero == slcBarberosAgenda || cita.cliente.email == email) && cita.fecha.getFullYear() == nuevaFecha.getFullYear() && cita.fecha.getMonth() == nuevaFecha.getMonth() && cita.fecha.getDate() == nuevaFecha.getDate() && cita.fecha.getHours() == nuevaFecha.getHours() && cita.fecha.getMinutes() == nuevaFecha.getMinutes())
 
     })
 
       if (existe) {
-        alert('¡Esta fecha y hora ya está agendada!');
         return 'No OK'
       } else {
 
@@ -183,7 +165,8 @@
           agendas.push(nuevoAgenda6)
           agendas.push(nuevoAgenda7)
         }
-        alert('Agendado');
         return 'OK'
       }
     }
+
+export {agendarCita};
